pyproject.toml
[tool.poetry]
name = "distributed_storage"
version = "0.1.0"
description = ""
authors = ["Yaroslav <y.kivaev@tinkoff.ru>"]
readme = "README.md"
packages = [{include = "distributed_storage"}]

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.56.0"
grpcio-tools = "^1.56.0"
google-api-python-client = "^2.93.0"
protobuf = "^4.23.4"
redis = "^4.6.0"
poetry = "^1.5.1"

[tool.poetry.scripts]
client = "distributed_storage.client:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
